Add files via upload
Service is implemented using two different third-party geocoding services. It is able
to have a primary service and additional backup services used for each request. For example, if the
primary geocoding service does not return a result or there is a network error when accessing the
service code will fall back to using a secondary service to complete the geocoding.
resolution.